🚚 rename component to a sensible name
